Fix GN generation for WinASAN build
This is a step toward a working Asan build with GN.
There are still a few missing pieces to land.

This patch is fixing an assert triggering due to courgette using an
harcoded toolchains.

----------------------
 ERROR at //build/config/sanitizers/sanitizers.gni:12:1: Assertion failed.
  assert(!using_sanitizer || is_clang,
  ^-----
  Sanitizers (is_*san) require setting is_clang = true in 'gn args'
----------------------

This patch is also fixing the compiler-rt runtime paths for a x64 bit
build. The port of WinASAN-64 is almost done and should part of the next
clang roll-deps.

R=thakis@chromium.org, dpranke@chromium.org
TBR=brettw@chromium.org
BUG=598761

Review-Url: https://codereview.chromium.org/2192833002
Cr-Original-Commit-Position: refs/heads/master@{#408857}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 364deb713883f28d2647c5ff160c1a6f7b7ef342